# backend/analytics-service/src/main/java/com/reoagms/analytics_service/common/model/ErrorResponse.java
package com.reoagms.analytics_service.common.model;

import lombok.Builder;
import lombok.Data;

import java.time.LocalDateTime;
import java.util.Map;

@Data
@Builder
public class ErrorResponse {
    private LocalDateTime timestamp;
    private int status;
    private String message;
    private String path;
    private Map<String, String> validationErrors;
}